feat: addressed review comms
